requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
trafilatura==1.8.0
html2text==2024.2.26
python-dotenv==1.0.1
faiss-cpu==1.12.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
numpy==1.26.4
scikit-learn==1.5.2
tenacity==8.5.0
pydantic==2.9.2
streamlit==1.38.0
anthropic==0.34.2
openai==1.51.2
ollama==0.3.3

